se realiza la rutina por dia de entrenamiento y tipo

+++ src/main/java/com/tallerwebi/dominio/ServicioMercadoPago.java
@@ -6,5 +6,5 @@ import com.tallerwebi.presentacion.DatosPreferencia;
 
 public interface ServicioMercadoPago {
 
-    DatosPreferencia crearPreferenciaPago(Integer total) throws MPException, MPApiException;
+    DatosPreferencia crearPreferenciaPago(double total) throws MPException, MPApiException;
 }
